# docs/swag/wallet/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Wallet API server",
        "title": "Wallet API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "ap11.0dev.cc",
    "basePath": "/",
    "paths": {
...
        "/transfer/:transfer_id/detail": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get detail transfer by transfer id and user id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transfer"
                ],
                "summary": "Get detail transfer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of transfer",
                        "name": "transfer_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully Get transfer detail",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid transfer ID or should greater 0",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User ID not found in context",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found - Transfer not found",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    }
                }
            }
        },
        "/transfer/claim": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Claim a pending transfer that was sent to the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transfer"
                ],
                "summary": "Claim transfer",
                "parameters": [
                    {
                        "description": "Transfer claim request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bitbucket_org_interdigital_aka-im-server-wallet_internal_domain.ClaimTransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully claimed transfer",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid transfer ID or already claimed",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User ID not found in context",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found - Transfer not found",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    }
                }
            }
        },
        "/transfer/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new transfer from authenticated user to another user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transfer"
                ],
                "summary": "Create transfer",
                "parameters": [
                    {
                        "description": "Transfer creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bitbucket_org_interdigital_aka-im-server-wallet_internal_domain.CreateTransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created transfer",
                        "schema": {
                            "$ref": "#/definitions/bitbucket_org_interdigital_aka-im-server-wallet_internal_domain.CreateTransferResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid transfer data or insufficient balance",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User ID not found in context",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found - Recipient user not found",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    }
                }
            }
        },
        "/transfer/refund": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Refund a transfer authentication user after expired",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transfer"
                ],
                "summary": "Self Refund Transfer",
                "parameters": [
                    {
                        "description": "Refund transfer request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bitbucket_org_interdigital_aka-im-server-wallet_internal_domain.RefundTransferReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully Refund transfer",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid transfer ID or should greater 0",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User ID not found in context",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found - Transfer not found",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apiresp.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errCode": {
                    "type": "integer"
                },
                "errDlt": {
                    "type": "string"
                },
                "errMsg": {
                    "type": "string"
                }
            }
        },
...
        "bitbucket_org_interdigital_aka-im-server-wallet_internal_domain.ClaimTransferRequest": {
            "type": "object",
            "properties": {
                "claimerUserID": {
                    "type": "string"
                },
                "operateBy": {
                    "type": "string"
                },
                "transferID": {
                    "type": "integer"
                }
            }
        },
...
        "bitbucket_org_interdigital_aka-im-server-wallet_internal_domain.CreateTransferRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "createdBy": {
                    "type": "string"
                },
                "fromUserID": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "toUserID": {
                    "type": "string"
                }
            }
        },
        "bitbucket_org_interdigital_aka-im-server-wallet_internal_domain.CreateTransferResponse": {
            "type": "object",
            "properties": {
                "transferID": {
                    "type": "integer"
                }
            }
        },
...
        "bitbucket_org_interdigital_aka-im-server-wallet_internal_domain.RefundTransferReq": {
            "type": "object",
            "properties": {
                "operatedBy": {
                    "type": "string"
                },
                "transferID": {
                    "type": "integer"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}